Take the group spinlock for the duration of the mail delivery.
al5_user_deliver needs to be called under the group lock to prevent the current user
to be unbind by the userspace while we are delivering the message.

There is a race condition where the mail might be delivered to a new user
or using a queue that is not initialized.

Signed-off-by: Kevin Grandemange <kevin.grandemange@allegrodvt.com>